pipeline: factor out queue message format
